Implement feature-based image selection/filtering:
Processors which need Workspace.image_from_page and image_from_segment
to deal with derived images sometimes need to be picky about what
features (i.e. AlternativeImage/@comments classes) they want to get
delivered (across the hierarchy). For example,
- segmentation might need non-binarized images, but depend on first
  running deskewing, which needs them binarized,
- deskewing needs unrotated images, but images deskewed on a higher
  level are ok,
- cropping needs unrotated, uncropped images, regardless of the steps
  already performed and annotation present.

To make all this possible, and rid of ambiguities existing in the current
scheme, here the API is extended and generalized to allow features
to be selected and filtered when calling these functions. That is,
* ``feature_selector`` takes a string encoded like AlternativeImage@comments
  (as a comma-separated list of feature classes) representing a logical
  AND mask of features to require from the chosen AlternativeImage
  and from the operations on it or on the original/parent,
* ``feature_filter`` takes a string encoded like AlternativeImage@comments
  (as a comma-separated list of feature classes) representing a logical
  OR mask of features to deny from the chosen AlternativeImage
  and from the operations on it or on the original/parent.

This also allows (and in fact requires) processors to correctly annotate
features (@comments) in a monotonically growing way (i.e. aggregating
all the operations actually performed to get the respective image).
Features are returned via a new ``features`` key in the xywh dictionary
of the second output position.

Documentation and consistency also vastly improves here.

Automated tests are still missing.