Switch from fchmod to chmod to allow dev on win
The method `os.fchmod` is only available on Unix system.
Therefore developing for `ocrd_utils` or any packages depending on
it is not possible on Windows.

As of Python 3.3 `os.chmod` also accepts a file descriptor.

Switching to `os.chmod` has the risk of masking raise conditions that
should be prohibited by `AtomicWriterPerms`, but only **on Windows**.
AFAIK running natively on Windows is not a goal of OCR-D (right now).