Change rabbitmq default credentials
Previously there were 'default-publisher' and 'default-consumer'. I
changed that to 'default'. Reason: For the rabbitmq no custom
Dockerfile is used. So users have to be added through rabbitmqctl via
docker-exec. This raises error if too many invocations happen. Maybe
misuse of the exec commands or bug in paramiko/docker-sdk is the cause.
But for now the easiest way to circumvent is to just reduce the
invocation count.