Add optional path_to_bin_dir to config and use it
It is asumed that on the targed machines all required ocrd-processors
are available in PATH to be called. At least for testing (and I think
for 'real' usage later too) it is much easier to create a venv with
ocrd to be used. Now with the path to this venv-bin-dir (dir where
all executables of venv can be found) it is possible to use it for
the processing-workers