Reuse parts of older impl for running a processor
It is not clear yet if the processing-broker should implement parts of
the webapi directly or if it should be usable on its own. In the latter
case the webapi would delegate to the processing-broker. Anyway I
changed the processing part to reuse code of the previous
processing-server PR (884) and it also uses the same processing
endpoint for now. This way it is currently easier for me to proceed but
maybe it has to be changed later again.